Add gitignore line parsers
This adds the intial line parsers for a gitignore file. For now, we are
not worrying about lines that include any '*' character. That will be
handled later.

Because vim's regular expression syntax is a little verbose and
different from the perl standard, the patterns are broken out into
variables so it can be a little more readable.